# Delayed symbol decoding