--- a/internal/frames.go
+++ b/internal/frames.go
@@ -121,7 +121,22 @@ func (e *TraceErr) Unwrap() error {
 
 // OrigError returns original wrapped error
 func (e *TraceErr) OrigError() error {
-	return e.Err
+	err := e.Err
+	// this is not an endless loop because I'm being
+	// paranoid, this is a safe protection against endless
+	// loops
+	for i := 0; i < MaxHops; i++ {
+		newerr, ok := err.(Error)
+		if !ok {
+			break
+		}
+		next := newerr.OrigError()
+		if next == nil || next == err {
+			break
+		}
+		err = next
+	}
+	return err
 }
 
 // GoString formats this trace object for use with
@@ -240,6 +255,34 @@ func UserMessage(err error) string {
 type UserMessager interface {
 	// UserMessage returns the user message associated with the error if any
 	UserMessage() string
+}
+
+// Error is an interface that helps to adapt usage of trace in the code
+// When applications define new error types, they can implement the interface
+//
+// Error handlers can use Unwrap() to retrieve error from the wrapper, or
+// errors.Is()/As() to compare it to another value.
+type Error interface {
+	error
+	ErrorWrapper
+	DebugReporter
+	UserMessager
+
+	// AddMessage adds formatted user-facing message
+	// to the error, depends on the implementation,
+	// usually works as fmt.Sprintf(formatArg, rest...)
+	// but implementations can choose another way, e.g. treat
+	// arguments as structured args
+	AddUserMessage(formatArg interface{}, rest ...interface{}) *TraceErr
+
+	// AddField adds additional field information to the error
+	AddField(key string, value interface{}) *TraceErr
+
+	// AddFields adds a map of additional fields to the error
+	AddFields(fields map[string]interface{}) *TraceErr
+
+	// GetFields returns any fields that have been added to the error
+	GetFields() map[string]interface{}
 }
 
 // Wrap returns a new instance of trace error at specified stack depth
@@ -444,3 +487,6 @@ Original Error: {{.OrigErrType}} {{.OrigErrMessage}}
 {{.Caught}}
 User Message: {{.UserMessage}}
 {{else}}User Message: {{.UserMessage}}{{end}}`
+
+// MaxHops is a max supported nested depth for errors
+const MaxHops = 50