This PR implements delayed decoding of debug symbols to lower the memory allocations and speed up error value creation in the basic wrapping scenario. The symbols are decoded when necessary (e.g. on DebugReport or upon being written to the wire).

Before the change
```
$ go test -bench=. -benchmem ./...
OK: 29 passed
goos: linux
goarch: amd64
pkg: github.com/gravitational/trace
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
BenchmarkStacktraceAllocs-16    	  642918	      1574 ns/op	     779 B/op	       6 allocs/op
```

After the change:
```
$ go test -bench=. -benchmem ./...
OK: 29 passed
goos: linux
goarch: amd64
pkg: github.com/gravitational/trace
cpu: Intel(R) Core(TM) i7-6700HQ CPU @ 2.60GHz
BenchmarkStacktraceAllocs-16    	 1348497	       829.1 ns/op	     400 B/op	       4 allocs/op
```

```
$ benchstat old.txt new.txt 
name                 old time/op    new time/op    delta
StacktraceAllocs-16    1.57µs ± 0%    0.83µs ± 0%   ~     (p=1.000 n=1+1)

name                 old alloc/op   new alloc/op   delta
StacktraceAllocs-16      779B ± 0%      400B ± 0%   ~     (p=1.000 n=1+1)

name                 old allocs/op  new allocs/op  delta
StacktraceAllocs-16      6.00 ± 0%      4.00 ± 0%   ~     (p=1.000 n=1+1)
```